pyproject.toml
[project]
name = "llmgrid"
version = "0.1.0"
description = "LLM agents navigating partially observable grid worlds with range-limited communication and artifacts."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.8",
  "pydantic-ai-slim[openai]>=0.0.16",
  "openai>=1.50.0",
  "httpx>=0.27.0",
  "typer>=0.12.0",
  "pyyaml>=6.0.1",
  "rich>=13.7.1",
  "Pillow>=10.3.0",
]

[project.scripts]
llmgrid-poc = "llmgrid.cli.poc_two_agents:main"
llmgrid-render-gif = "llmgrid.cli.render_gif:main"

[tool.uv]
dev-dependencies = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.9",
]
